uint64 timestamp

+++ server.go
@@ -79,7 +79,7 @@ func mapMetrics(data map[string]interface{}) []*Metric {
 	for key, value := range data {
 		name, ok := rootMetrics[key]
 		if ok {
-			metrics = append(metrics, NewMetric(host, name, timestamp, value, nil))
+			metrics = append(metrics, NewMetric(host, name, uint64(timestamp), value, nil))
 		} else {
 			fmt.Println(key)
 		}
